Fix bug where multiple blobs could name the same successor.
Because the blob-finding scoring function is symmetric in time, we can compare
scores from found-to-lost just as lost-to-found (what we supposedly calculate).
I use a dictionary that keeps track of what the highest scoring connection was
and to what then convert it back to a simple list of 'edges' before passing it
to a split-off `join_segments` function.