Update tests to check for trying to pass one blob trace to the scorer
Now we raise an exception, `OneGoodBlobException` when attempting to do so.
Callers can handle this appropriately, likely by not attempting to generate
a scoring method because there is nothing to join that would need scoring.

Also updated `where.py` to have a default folder it checks for data sets in so
the simple YYYYMMDD_HHmmss identifier can be passed.